refactor(hiring-process): move componentes para dentro de pages/hiring-process
@leonardogomesp

34
