+++ src/components/exercises-list/index.js
@@ -1,5 +1,5 @@
 import './style.css'
-import { ToggleRow } from '../togglerow'
+import { ToggleRow } from '../toggle-row'
 import { useEffect, useState } from 'react'
 import { client } from '../../service'
