+++ src/pages/hiring-process/forms/hiring-process/index.js
@@ -1,8 +1,8 @@
 import { useState } from 'react'
-import { InputDate } from '../../inputs/date'
-import { InputText } from '../../inputs/text'
-import { client } from '../../../service'
-import Button from '../../button'
+import { InputDate } from '../../../../components/inputs/date'
+import { InputText } from '../../../../components/inputs/text'
+import { client } from '../../../../service'
+import Button from '../../../../components/button'
 
 export const HiringProcessForm = ({ method = 'POST', id = '', callback = () => { } }) => {
   const [hiringProcess, setHiringProcess] = useState({
