+++ src/pages/hiring-process/index.js
@@ -1,7 +1,7 @@
 import Button from '../../components/button'
 import { Modal } from '../../components/modal'
-import { HiringProcessForm } from '../../components/forms/hiring-process'
-import { ProcessList } from '../../components/hiring-process-list'
+import { HiringProcessForm } from './forms/hiring-process'
+import { ProcessList } from '../hiring-process/components/hiring-process-list'
 import './style.css'
 import { faPlus } from '@fortawesome/free-solid-svg-icons'
 import { useEffect, useState } from 'react'
